+++ src/MapEditor.Avalonia/Views/Items/ItemsWindow.axaml.cs
@@ -1,0 +1,9 @@
+﻿using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace MapEditor.Avalonia.Views.Items;
+
+public partial class ItemsWindow : Window
+{
+    public ItemsWindow() => AvaloniaXamlLoader.Load(this);
+}